pyproject.toml
[tool.poetry]
name = "tic-tac-gpt"
version = "0.1.0"
description = ""
authors = ["andylolu2 <andylolu24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.2.0"
ml-tools = {path = "../ml-tools", develop = true}
absl-py = "^2.1.0"
transformer-lens = "^1.14.0"
lightning = "^2.2.0.post0"
pytest = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
